Remove check for map open
Checks if map is visible is already built into the controls, there is no need to check this as well. Indicators cannot rander on map, so they do not need it as well.

What is the benefit of this change? When closing the map, namelist and compass show immediately., meaning they do not pop in and may distract.